refactor: consolidate filtering logic
